Related to #87 and #85.
This should patch part of the issue where it couldn't map MeSH tags to RxNorms due to changes of tagging on the MEDLINE side (D codes to C codes).

The other 1/2 of the fix will come from Rich when rolls-back removal of PharmacoAction mapping piece implemented in the last release.